requirements.txt
fastapi
uvicorn
python-multipart
langchain==0.3.25
langchain-ollama
langchain-community
langchain-core>=0.3.0
langchain-text-splitters
chromadb
tree-sitter
tree-sitter-languages
PyMuPDF
pyyaml
streamlit
requests
langchain-groq
python-dotenv
slowapi